fix: fix possibly ambiguous or broken rule
This rule was intended to apply to dependencies that are sourced from
npm *and* whose names match the specified patterns. It's not at all
clear that it was behaving that way. It may have been matching
dependencies that are sourced from npm *or* whose names match the
specified patterns, which would give unintended behaviour for most
npm dependencies.

The Renovate docs are very vague about whether matchers in a
packageRule are supposed to form an "and" relationship or an "or"
relationship. I intended an "and" relationship but in practice
Renovate seems to implement an "or" relationship. I don't know if
this is by design. There doesn't seem to be a way to robustly
form an "and" relationship. Nesting packageRules might work but
let's not risk it unnecessarily.